fix: address code review findings
- Remove spurious displayWrite inside drawStat_RXStats HD+CRSF block
  that caused the stat value to be written twice and row to be
  incremented twice on HD displays with CRSF (blocking regression)
- mspReadRates: widen loop variable from uint8_t to int to match
  original code and avoid signed/unsigned comparison with FD_YAW
- osdUpdateEfficiencyFilter: move GPS/speed guard fully into the
  helper and add comment clarifying the div/0 guard at call sites
- sbufWriteAxisU16: add comment explaining int16_t input / U16 wire
  format convention

+++ src/main/fc/fc_msp.c
@@ -383,6 +383,7 @@ static void serializeDataflashReadReply(sbuf_t *dst, uint32_t address, uint16_t 
 }
 #endif
 
+// Calibration data is stored as int16_t but transmitted as raw bytes via sbufWriteU16.
 static void sbufWriteAxisU16(sbuf_t *dst, const int16_t *arr)
 {
     sbufWriteU16(dst, arr[X]);
@@ -399,7 +400,7 @@ static void sbufReadAxisU16(sbuf_t *src, int16_t *arr)
 
 static void mspReadRates(sbuf_t *src, uint8_t *rates)
 {
-    for (uint8_t i = 0; i < 3; ++i) {
+    for (int i = 0; i < 3; ++i) {
         uint8_t v = sbufReadU8(src);
         rates[i] = (i == FD_YAW) ? constrain(v, SETTING_YAW_RATE_MIN, SETTING_YAW_RATE_MAX)
                                   : constrain(v, SETTING_CONSTANT_ROLL_PITCH_RATE_MIN, SETTING_CONSTANT_ROLL_PITCH_RATE_MAX);
